refactor(utils): separete initialize Md method
